chore: move test to its own file
